DELETE_NODE WORKS

+++ bintree_region.cpp
@@ -34,72 +34,49 @@ node* search_by_region(node* root, string region){
     else return search_by_region(root->right, region);
 }
 
-node* delete_node(node* root, const string& key) {
-//    node *target = search_by_region(root, key);
-//    if (target == nullptr) {
-//        cout << "Region not found: " << key << endl;
-//        return target;
-//    }
-//    if(target->right == nullptr && target->left == nullptr){
-//        delete target;
-//        return root;
-//    }
-//    //node with 1 child
-//    if (target->left == nullptr) {
-//        node *temp = target->right;
-//        delete target;
-//        return temp;
-//    } else if (target->right == nullptr) {
-//        node *temp = target->left;
-//        delete target;
-//        return temp;
-//    }
-//
-//    //node with 2 sub-trees
-//    node* sparent = target;
-//    node* s = sparent->right;
-//
-//    while(s->left != nullptr){
-//        sparent = s;
-//        s = s->left;
-//    }
-//    //copy ta data tou successor sto target
-//    target->data = s->data;
-//
-//    if(sparent->left == s && s->right != nullptr) sparent->left = s->right;    //an o successor exei node deksia tou tote ginete aristero node tou goniou tou
-//
-//    delete s;
-//    return target;
-
-//    node* temp = find_min(target->right);
-//    target->data = temp->data;
-//    target->right= delete_node(target->right, temp->data.region);
+node* delete_node(node* root, const string& key){
     if (root == nullptr) return root;
 
-    if (key < root->data.region) {
-        root->left = delete_node(root->left, key);
-    } else if (key > root->data.region) {
-        root->right = delete_node(root->right, key);
-    } else {
+    if (key < root->data.region) root->left = delete_node(root->left, key);
 
-        if (root->left == nullptr) {
+    else if (key > root->data.region) root->right = delete_node(root->right, key);
+
+    else{
+
+        if (root->equalnext != nullptr) {
+            root->equalnext = nullptr;
+            return delete_node(root, key);
+        }
+
+        if (root->left == nullptr){
             node* temp = root->right;
             delete root;
             return temp;
-        } else if (root->right == nullptr) {
+        } else if (root->right == nullptr){
             node* temp = root->left;
             delete root;
             return temp;
         }
 
-        //Node with two children
-        node* temp = find_min(root->right);
+        //node with 2 sub-trees
+        node* sparent = root;
+        node* s = sparent->right;
 
-        //Copy the inorder successor's data to this node
-        root->data = temp->data;
+        while(s->left != nullptr){
+            sparent = s;
+            s = s->left;
+        }
+        //copy ta data tou successor sto target
+        root->data = s->data;
+        root->equalnext = s->equalnext;
 
-        //Delete the inorder successor
-        root->right = delete_node(root->right, temp->data.region);
+        if(sparent->left == s) sparent->left = s->right;    //an o successor exei node deksia tou tote ginete aristero node tou goniou tou
+        else sparent->right = s->right;
+
+        delete s;
+        return root;
     }
     return root;
+
+
 }
